Support inline pagination on tree listings
GitLab 10.x changed per-request limits on the number of items returned
from the tree listing API, capping at 100 per page with no evident way
to change that as a system-wide configuration. To work around this in
O'Reilly Atlas, we added simple pagination to continue making requests
to the API until all file in a directory are loaded.

This required one change to the Backbone-GitLab API wrapper, to memoize
the per_page option so it's not reset on subsequent requests during
paginated loading. Backport this change to the backbone-gitlab package.

https://intranet.oreilly.com/jira/browse/ATLAS-169
https://intranet.oreilly.com/jira/browse/ATLAS-147